% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dataset_delaunay_generate}
\alias{dataset_delaunay_generate}
\title{Generate Delaunay triangulation network data}
\usage{
dataset_delaunay_generate(
  classes = 0L,
  n_samples_in_class = 1000L,
  n_nodes = 7L,
  support_low = 0,
  support_high = 10,
  drift_amount = 1,
  one_hot_labels = TRUE,
  support = NULL,
  seed = NULL,
  return_type = "numpy"
)
}
\arguments{
\item{classes}{classes indices of the classes to load (integer, or list of integers
between 0 and 20)}

\item{n_samples_in_class}{number of generated samples per class}

\item{n_nodes}{n_nodes number of nodes in a graph}

\item{support_low}{support_low lower bound of the uniform distribution from which the
support is generated}

\item{support_high}{support_high upper bound of the uniform distribution from which the
support is generated}

\item{drift_amount}{drift_amount coefficient to control the amount of change between classes}

\item{one_hot_labels}{one_hot_labels one-hot encode dataset labels}

\item{support}{support custom support to use instead of generating it randomly}

\item{seed}{seed random numpy seed}

\item{return_type}{Data format to return data in. One of either "list", or "tidygraph"}
}
\value{
\itemize{
\item if \code{return_type='list'}, the adjacency matrix, node features, and
an array containing labels, in a list
\item if \code{return_type='tidygraph'}, a tidygraph object, with node features and labels as
node data
}
}
\description{
Generates a dataset of Delaunay triangulations as described by \href{https://arxiv.org/abs/1706.06941}{Zambon et al. (2017)}.
Node attributes are the 2D coordinates of the points.
Two nodes are connected if they share an edge in the Delaunay triangulation.
Labels represent the class of the graph (0 to 20, each class index i
represent the "difficulty" of the classification problem 0 v. i. In other
words, the higher the class index, the more similar the class is to class 0).
}
