% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dataset_graph_mnist}
\alias{dataset_graph_mnist}
\title{MNIST Graph Dataset}
\usage{
dataset_graph_mnist(k = 8L, noise_level = 0)
}
\arguments{
\item{k}{int, number of neighbours for each node}

\item{noise_level}{fraction of edges to flip (from 0 to 1 and vice versa)}

\item{return_type}{Data format to return data in. One of either "list", or "tidygraph"}
}
\value{
\itemize{
\item X_train, y_train: training node features and labels; - X_val, y_val: validation node features and labels; - X_test, y_test: test node features and labels; - A: adjacency matrix of the grid;
}
}
\description{
Loads the MNIST dataset and a K-NN graph to perform graph signal classification, as described by \href{https://arxiv.org/abs/1606.09375}{Defferrard et al. (2016)}.
The K-NN graph is statically determined from a regular grid of pixels using
the 2d coordinates. The node features of each graph are the MNIST digits vectorized and rescaled
to \link{0, 1}.
Two nodes are connected if they are neighbours according to the K-NN graph.
Labels are the MNIST class associated to each sample.
}
