% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dataset_graphsage}
\alias{dataset_graphsage}
\title{GraphSage Datasets}
\usage{
dataset_graphsage(dataset_name, max_degree = -1L, normalize_features = TRUE)
}
\arguments{
\item{dataset_name}{dataset_name name of the dataset to load (\code{'ppi'}, or \code{'reddit'})}

\item{max_degree}{max_degree if positive, subsample edges so that each node has
the specified maximum degree.}

\item{normalize_features}{normalize_features if TRUE, the node features are normalized}

\item{return_type}{Data format to return data in. One of either "list", or "tidygraph"}
}
\value{
Either a list with 6 elements (containing - Adjacency matrix, Node features,
Labels, and 3 binary masks for train, validation, and test splits), or a \code{tbl_graph}
object with the Node features, Labels, and 3 binary masks as node attributes).
}
\description{
Loads one of the datasets (PPI or Reddit) used in \href{https://arxiv.org/abs/1706.02216}{Hamilton & Ying (2017)}.
The PPI dataset (originally \href{https://www.ncbi.nlm.nih.gov/pubmed/16381927}{Stark et al. (2006)})
for inductive node classification uses positional gene sets, motif gene sets
and immunological signatures as features and gene ontology sets as labels.
The Reddit dataset consists of a graph made of Reddit posts in the month of
September, 2014. The label for each node is the community that a
post belongs to. The graph is built by sampling 50 large communities and
two nodes are connected if the same user commented on both. Node features
are obtained by concatenating the average GloVe CommonCrawl vectors of
the title and comments, the post's score and the number of comments.
The train, test, and validation splits are returned as binary masks.
:param max_degree: int, if positive, subsample edges so that each node has
the specified maximum degree.
:param normalize_features: if TRUE, the node features are normalized;
:
}
