% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dataset_tud}
\alias{dataset_tud}
\title{TU Dortmund Benchmark Dataset}
\usage{
dataset_tud(dataset_name, clean = FALSE)
}
\arguments{
\item{dataset_name}{name of the dataset to load (see \code{spektral_datasets$tud$AVAILABLE_DATASETS}).}

\item{clean}{if TRUE, return a version of the dataset with no isomorphic graphs.}
}
\value{
A list of:
\itemize{
\item a list of adjacency matrices;
\item a list of node feature matrices;
\item an array containing the one-hot encoded targets.
}
}
\description{
Loads one of the Benchmark Data Sets for Graph Kernels from TU Dortmund
(\href{https://ls11-www.cs.tu-dortmund.de/staff/morris/graphkerneldatasets}{link}).
The node features are computed by concatenating the following features for
each node: - node attributes, if available, normalized as specified in \code{normalize_features};
\itemize{
\item clustering coefficient, normalized with z-score;
\item node degrees, normalized as specified in \code{normalize_features};
\item node labels, if available, one-hot encoded.
}
}
