% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_spektral}
\alias{install_spektral}
\title{Install spektral keras extension library in Python}
\usage{
install_spektral(version = NULL, ..., restart_session = TRUE)
}
\arguments{
\item{version}{WHat version of spektral to install? Default is to install latest version
available in pip}

\item{...}{Further arguments passed to \code{\link[reticulate]{py_install}}}

\item{restart_session}{Should the session be restarted after installation?}
}
\description{
Install spektral keras extension library in Python
}
