% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_conv.R
\name{layer_diffuse_features}
\alias{layer_diffuse_features}
\title{DiffuseFeatures}
\usage{
layer_diffuse_features(
  object,
  num_diffusion_steps,
  kernel_initializer,
  kernel_regularizer,
  kernel_constraint,
  ...
)
}
\arguments{
\item{num_diffusion_steps}{How many diffusion steps to consider. \(K\) in paper.}

\item{kernel_initializer}{initializer for the weights}

\item{kernel_regularizer}{regularization applied to the kernel vectors}

\item{kernel_constraint}{constraint applied to the kernel vectors}
}
\description{
\loadmathjax Utility layer calculating a single channel of the
diffusional convolution.

Procedure is based on https://arxiv.org/abs/1707.01926

\strong{Input}
\itemize{
\item Node features of shape \verb{([batch], N, F)};
\item Normalized adjacency or attention coef. matrix \mjeqn{\hat A }{} of shape
\verb{([batch], N, N)}; Use DiffusionConvolution.preprocess to normalize.
}

\strong{Output}
\itemize{
\item Node features with the same shape as the input, but with the last
dimension changed to \mjeqn{1}{}.
}
}
