% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_conv.R
\name{layer_gated_graph_conv}
\alias{layer_gated_graph_conv}
\title{GatedGraphConv}
\usage{
layer_gated_graph_conv(
  object,
  channels,
  n_layers,
  activation = NULL,
  use_bias = TRUE,
  kernel_initializer = "glorot_uniform",
  bias_initializer = "zeros",
  kernel_regularizer = NULL,
  bias_regularizer = NULL,
  activity_regularizer = NULL,
  kernel_constraint = NULL,
  bias_constraint = NULL,
  ...
)
}
\arguments{
\item{channels}{integer, number of output channels}

\item{n_layers}{integer, number of iterations with the GRU cell}

\item{activation}{activation function to use}

\item{use_bias}{bool, add a bias vector to the output}

\item{kernel_initializer}{initializer for the weights}

\item{bias_initializer}{initializer for the bias vector}

\item{kernel_regularizer}{regularization applied to the weights}

\item{bias_regularizer}{regularization applied to the bias vector}

\item{activity_regularizer}{regularization applied to the output}

\item{kernel_constraint}{constraint applied to the weights}

\item{bias_constraint}{constraint applied to the bias vector.}
}
\description{
\loadmathjax
A gated graph convolutional layer as presented by
\href{https://arxiv.org/abs/1511.05493}{Li et al. (2018)}.

\strong{Mode}: single, disjoint.

\strong{This layer expects a sparse adjacency matrix.}

This layer repeatedly applies a GRU cell \mjeqn{L}{} times to the node attributes
\mjdeqn{ \begin{align} & h^{(0)} _ i = X_i \| \mathbf{0} \\ & m^{(l)} _ i = \sum\limits_{j \in \mathcal{N}(i)} h^{(l - 1)} _ j W \\ & h^{(l)} _ i = \textrm{GRU} \left(m^{(l)} _ i, h^{(l - 1)} _ i \right) \\ & Z_i = h^{(L)} _ i \end{align} }{}
where \mjeqn{\textrm{GRU}}{} is the GRU cell.

\strong{Input}
\itemize{
\item Node features of shape \verb{(N, F)}; note that \code{F} must be smaller or equal
than \code{channels}.
\item Binary adjacency matrix of shape \verb{(N, N)}.
}

\strong{Output}
\itemize{
\item Node features with the same shape of the input, but the last dimension
changed to \code{channels}.
}
}
