% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_pool.R
\name{layer_global_avg_pool}
\alias{layer_global_avg_pool}
\title{GlobalAvgPool}
\usage{
layer_global_avg_pool(object, ...)
}
\description{
\loadmathjax
An average pooling layer. Pools a graph by computing the average of its node
features.

\strong{Mode}: single, disjoint, mixed, batch.

\strong{Input}
\itemize{
\item Node features of shape \verb{([batch], N, F)};
\item Graph IDs of shape \verb{(N, )} (only in disjoint mode);
}

\strong{Output}
\itemize{
\item Pooled node features of shape \verb{(batch, F)} (if single mode, shape will
be \verb{(1, F)}).
}

\strong{Arguments}

None.
}
