% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_conv.R
\name{layer_tag_conv}
\alias{layer_tag_conv}
\title{TAGConv}
\usage{
layer_tag_conv(
  object,
  channels,
  K = 3,
  activation = NULL,
  use_bias = TRUE,
  kernel_initializer = "glorot_uniform",
  bias_initializer = "zeros",
  kernel_regularizer = NULL,
  bias_regularizer = NULL,
  activity_regularizer = NULL,
  kernel_constraint = NULL,
  bias_constraint = NULL,
  ...
)
}
\arguments{
\item{channels}{integer, number of output channels}

\item{K}{the order of the layer (i.e., the layer will consider a K-hop
neighbourhood for each node)}

\item{activation}{activation function to use}

\item{use_bias}{bool, add a bias vector to the output}

\item{kernel_initializer}{initializer for the weights}

\item{bias_initializer}{initializer for the bias vector}

\item{kernel_regularizer}{regularization applied to the weights}

\item{bias_regularizer}{regularization applied to the bias vector}

\item{activity_regularizer}{regularization applied to the output}

\item{kernel_constraint}{constraint applied to the weights}

\item{bias_constraint}{constraint applied to the bias vector.}
}
\description{
\loadmathjax
A Topology Adaptive Graph Convolutional layer (TAG) as presented by
\href{https://arxiv.org/abs/1710.10370}{Du et al. (2017)}.

\strong{Mode}: single, disjoint.

\strong{This layer expects a sparse adjacency matrix.}

This layer computes:
\mjdeqn{ Z = \sum\limits_{k=0}^{K} D^{-1/2}A^kD^{-1/2}XW^{(k)} }{}

\strong{Input}
\itemize{
\item Node features of shape \verb{(N, F)};
\item Binary adjacency matrix of shape \verb{(N, N)}.
}

\strong{Output}
\itemize{
\item Node features with the same shape of the input, but the last dimension
changed to \code{channels}.
}
}
