% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_arma_conv}
\alias{preprocess_arma_conv}
\title{Preprocess Adjacency Matrix for use with layer_arma_conv}
\usage{
preprocess_arma_conv(A)
}
\arguments{
\item{A}{An Adjacency matrix (can be dense or sparse)}
}
\description{
This utility function can be used to  preprocess a network adjacency matrix
into an object that can be used to represent the network in the \code{\link{layer_arma_conv}} layer.
Internally it does this:\cr
\code{fltr=\link{utils_normalized_laplacian}(A,symmetric=True)}\cr
\code{fltr=\link{utils_rescale_laplacian}(fltr,lmax=2)}\cr
\code{fltr}\cr
}
