% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_chebyshev}
\alias{preprocess_chebyshev}
\title{Calculate Chebyshev polynomials from an adjacency matrix}
\usage{
preprocess_chebyshev(A, densify = FALSE)
}
\arguments{
\item{A}{Adjacency matrix as a sparse matrix (\code{\link[Matrix]{dgRMatrix}})}

\item{densify}{Should the result be returned as a dense matrix (default is sparse)?}
}
\value{
Chebyshev polynomials as a sparse matrix (or a dense matrix if \code{densify} is \code{TRUE})
}
\description{
This is used to preprocess and adjacency matrix so it can be used directly
in a Chebyshev graph convolutional layer, as implement by \code{\link{layer_cheb_conv}}
}
