% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_graph_conv_skip}
\alias{preprocess_graph_conv_skip}
\title{Preprocess Adjacency Matrix for use with layer_graph_conv_skip}
\usage{
preprocess_graph_conv_skip(A)
}
\arguments{
\item{A}{An Adjacency matrix (can be dense or sparse)}
}
\description{
This utility function can be used to  preprocess a network adjacency matrix
into an object that can be used to represent the network in the \code{\link{layer_graph_conv_skip}} layer.
Internally it does this:\cr
\code{\link{utils_normalized_adjacency}(A)}\cr
}
