% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_laplacian_mod}
\alias{preprocess_laplacian_mod}
\title{Calculate modified Laplacian on an adjacency matrix}
\usage{
preprocess_laplacian_mod(A, densify = FALSE)
}
\arguments{
\item{A}{Adjacency matrix as a sparse matrix (\code{\link[Matrix]{dgRMatrix}})}

\item{densify}{Should the Laplacian be returned as a dense matrix (default is sparse)?}
}
\value{
The Laplacian as a sparse matrix (or a dense matrix if \code{densify} is \code{TRUE})
}
\description{
This is used to preprocess and adjacency matrix so it can be used directly
in a graph convolutional network (GCN), as implement by \code{\link{layer_graph_conv}}
}
