% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_localpooling_filter}
\alias{utils_localpooling_filter}
\title{Compute Graph Filter}
\usage{
utils_localpooling_filter(A, symmetric = TRUE)
}
\arguments{
\item{A}{array or sparse matrix with rank 2 or 3}

\item{symmetric}{boolean, whether to normalize the matrix as \mjeqn{D^{-\frac{1}{2}}AD^{-\frac{1}{2}}}{D^(-1/2)AD^(-1/2)} or as \mjeqn{D^{-1}A}{D^(-1)A}}
}
\value{
array or sparse matrix with rank 2 or 3, same as A
}
\description{
\loadmathjax
Computes the graph filter described in \href{https://arxiv.org/abs/1609.02907}{Kipf & Welling (2017)}.
}
