% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_normalized_adjacency}
\alias{utils_normalized_adjacency}
\title{Normalize Adjacency Matrix}
\usage{
utils_normalized_adjacency(A, symmetric = TRUE)
}
\arguments{
\item{A}{rank 2 array or sparse matrix}

\item{symmetric}{boolean, compute symmetric normalization}
}
\value{
the normalized adjacency matrix.
}
\description{
\loadmathjax
Normalizes the given adjacency matrix using the degree matrix as either
\mjeqn{D^{-1}A}{D^(-1)A} or \mjeqn{D^{-1/2}AD^{-1/2}}{D^(-1/2)AD^(-1/2)} (symmetric normalization).
}
