% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_normalized_laplacian}
\alias{utils_normalized_laplacian}
\title{Compute normalized Laplacian of an adjacency matrix}
\usage{
utils_normalized_laplacian(A, symmetric = TRUE)
}
\arguments{
\item{A}{rank 2 array or sparse matrix;}

\item{symmetric}{boolean, compute symmetric normalization}
}
\value{
the normalized Laplacian.
}
\description{
\loadmathjax
Computes a normalized Laplacian of the given adjacency matrix as
\mjeqn{I - D^{-1}A}{I - D^(-1)A} or \mjeqn{I - D^{-1/2}AD^{-1/2}}{I - D^(-1/2)AD^(-1/2)} (symmetric normalization).
}
