% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_rescale_laplacian}
\alias{utils_rescale_laplacian}
\title{Rescale a Laplacian}
\usage{
utils_rescale_laplacian(L, lmax = NULL)
}
\arguments{
\item{L}{rank 2 array or sparse matrix}

\item{lmax}{if NULL, compute largest eigenvalue with scipy.linalg.eisgh.
If the eigendecomposition fails, lmax is set to 2 automatically.
If scalar, use this value as largest eignevalue when rescaling.}
}
\value{
the rescaled Laplacian
}
\description{
\loadmathjax
Rescales the Laplacian eigenvalues in \link{-1,1}, using lmax as largest eigenvalue.
}
